+++ src/Types.ts
@@ -1,4 +1,100 @@
 /**
+ * Represents the configuration object for the application.
+ */
+export type config = {
+	/**
+	 * An object containing permission settings.
+	 */
+	Permission: {
+		bypassPermission: string;
+		useTxAdmin: boolean;
+	};
+	/**
+	 * The Discord webhook URL for logging.
+	 */
+	DiscordWebhook: string;
+	/**
+	 * An object containing module settings.
+	 */
+	Modules: {
+		[key: string]: {
+			/**
+			 * Obviously there is more data being hold by a module,
+			 * but as of now there is no need to represent it here.
+			 */
+			enabled: boolean;
+		};
+	};
+	/**
+	 * An array of blacklisted weapon names.
+	 */
+	BlacklistedWeapons: string[];
+	/**
+	 * An array of blacklisted event names.
+	 */
+	BlacklistedEvents: {
+		[key: string]: string[];
+	};
+	/**
+	 * An array of illegal model names.
+	 */
+	IllegalModels: string[];
+};
+/**
+ * Represents a set of deferral actions.
+ */
+export type Deferrals = {
+	/**
+	 * Defers the execution of the script.
+	 */
+	defer: () => void;
+	/**
+	 * Updates the deferral message.
+	 * @param message - The new message to display.
+	 */
+	update: (message: string) => void;
+	/**
+	 * Presents a card to the user.
+	 * @param card - The card object or string representation.
+	 * @param cb - Optional callback function to handle the card data and raw data.
+	 */
+	presentCard: (card: object | string, cb?: (data: object, rawData: string) => void) => void;
+	/**
+	 * Marks the deferral as done.
+	 * @param failureReason - Optional reason for deferral failure.
+	 */
+	done: (failureReason?: string) => void;
+};
+/**
+ * Represents a deferrals config object.
+ */
+export type DeferralsObject = {
+	enabled: boolean;
+	rejectionMsg: string;
+};
+/**
+ * Represents a help parameter for a command suggestion.
+ */
+export type Parameter = {
+	name: string;
+	help: string;
+};
+/**
+ * Represents release version data from github.
+ */
+export type Release = {
+	id: number;
+	name: string;
+	prerelease: boolean;
+};
+/**
+ * Represents a discord webhook payload object.
+ */
+export type WebhookPayload = {
+	username: string;
+	embeds?: object;
+};
+/**
  * Represents an event that occurs when a weapon causes damage.
  */
 export type WeaponDamageEvent = {
@@ -36,7 +132,6 @@ export type WeaponDamageEvent = {
 	weaponType: number;
 	willKill: boolean;
 };
-
 /**
  * Represents an ExplosionEvent.
  */
@@ -74,7 +169,6 @@ export type ExplosionEvent = {
 	f218: number;
 	f216: number;
 };
-
 /**
  * Represents the data structure for a FireEvent.
  */
@@ -95,7 +189,6 @@ export type FireEvent = {
 	weaponHash: number;
 	fireId: number;
 };
-
 /**
  * Represents an event for giving a weapon to a ped.
  */
@@ -121,7 +214,6 @@ export type GiveWeaponEvent = {
 	 */
 	givenAsPickup: boolean;
 };
-
 /**
  * Represents an event for removing a weapon from a ped.
  */
@@ -135,12 +227,10 @@ export type RemoveWeaponEvent = {
 	 */
 	weaponType: number;
 };
-
 /**
  * Represents an event that removes all weapons from a ped.
  */
 export type RemoveAllWeaponsEvent = { pedId: number };
-
 /**
  * Represents the event data for starting a projectile.
  */
@@ -175,7 +265,6 @@ export type StartProjectileEvent = {
 	unkZ8: number;
 	weaponHash: number;
 };
-
 /**
  * Represents a PtFxEvent.
  */
@@ -204,7 +293,6 @@ export type PtFxEvent = {
 	rotZ: number;
 	scale: number;
 };
-
 /**
  * Represents an player enetered/left scope event.
  */
@@ -212,7 +300,6 @@ export type PlayerScopeEvent = {
 	for: number;
 	player: number;
 };
-
 /**
  * Represents an event for txAdmin authentication.
  */
