+++ src/commands/LoadModuleCommand.ts
@@ -1,7 +1,7 @@
 import { ModuleLoader } from "../core/ModuleLoader";
 import { Command } from "../core/Command";
 import { Logger } from "../core/logger/Logger";
-import { Parameter } from "../types/ParameterType";
+import { Parameter } from "../Types";
 
 export class LoadModuleCommand extends Command {
 	/**
