--- src/types/ParameterType.ts
@@ -1,7 +1,0 @@
-/**
- * Represents a help parameter for a command suggestion.
- */
-export type Parameter = {
-	name: string;
-	help: string;
-};
