--- src/types/ReleaseType.ts
@@ -1,8 +1,0 @@
-/**
- * Represents release version data from github.
- */
-export type Release = {
-	id: number;
-	name: string;
-	prerelease: boolean;
-};
