--- src/types/WebhookPayloadType.ts
@@ -1,4 +1,0 @@
-export type WebhookPayload = {
-	username: string;
-	embeds?: object;
-};
